// NgoPortal/src/app/model/donation.ts
export class Donation {
    constructor(){}
    id?: number | null | undefined;
    eventId?: number |null |undefined;
    eventName?: string |null | undefined;
    userId?: number |null |undefined;
    amount?: number | null | undefined;
    donationDate?: Date | null | undefined;
    reocurring?: boolean | null | undefined;
    firstName?: string |null | undefined;
    lastName?: string |null | undefined;
    phone?: string |null | undefined;
    address?: string |null | undefined;
    city?: string |null | undefined;
    state?: string |null | undefined;
    zipcode?: string |null | undefined;
    country?: string |null | undefined;
    email?:string |null | undefined;
  }